Using cd ... || exit in case cd fails.
cd can fail for a variety of reasons: misspelled paths, missing directories, missing permissions, broken symlinks, etc. 

If/when it does, the script will keep going and do all its operations in the wrong directory. This can be messy, especially if the operations involve creating or deleting a lot of files.